Orchestrator/requirements.txt
fastapi
uvicorn
requests
pydantic
httpx
python-dotenv
google-generativeai